Optimize Data type DB

+++ BE/oem-ev-warranty-management-system/src/main/java/com/swp391/warrantymanagement/repository/ServiceHistoryRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServiceHistoryRepository extends JpaRepository<ServiceHistory, Integer> {
+public interface ServiceHistoryRepository extends JpaRepository<ServiceHistory, Long> {
     // Spring Boot đã tự động sinh các phương thức CRUD cơ bản
+    // Spring Boot đã tự động sinh các phương thức CRUD cơ bản
+    // Này là Query method, tự động sinh câu truy vấn dựa trên tên phương thức
 }
